feat(scheduler): 카풀 종료 후 스케줄러 분리
- 카풀 종료 후 3시간 뒤 알람 전송, 이용내역 저장
- 카풀 종료 후 7일 뒤 firestore 삭제
- HistoryEntity 인위적 식별자 삭제, carId를 PK로

+++ inhaCarPool_SERVER/src/main/java/com/example/inhaCarpool/service/CarpoolService.java
@@ -1,6 +1,7 @@
 package com.example.inhaCarpool.service;
 
 import com.example.inhaCarpool.dto.CarpoolResponseDTO;
+import com.example.inhaCarpool.dto.CarpoolShortDTO;
 import com.example.inhaCarpool.dto.HistoryRequestDTO;
 import com.google.api.core.ApiFuture;
 import com.google.cloud.firestore.*;
@@ -31,13 +32,16 @@ public class CarpoolService {
     @SuppressWarnings("unchecked")
     public void scheduleCarpools() {
 
+        System.out.println("스케줄러 실행 중");
+
         Firestore firestore = FirestoreClient.getFirestore(); // Firestore 인스턴스 생성
 
-        ScheduledExecutorService scheduler = Executors.newScheduledThreadPool(1); // 스케줄러 생성
+        ScheduledExecutorService scheduler = Executors.newScheduledThreadPool(2); // 스케줄러 생성
 
+
+        // 출발시간으로부터 3시간 이상이 지난 카풀 알람 보내기(topic 구독 해제하므로 1회만 보내짐) -> history 이용내역에 저장
         scheduler.scheduleAtFixedRate(() -> {
             try {
-                System.out.println("스케줄러 실행 중");
 
                 ApiFuture<QuerySnapshot> future = firestore.collection(COLLECTION_NAME).get();
 
@@ -64,9 +68,6 @@ public class CarpoolService {
                                 member2 = memberArray.get(1).toString();
                                 member3 = memberArray.get(2).toString();
                             }
-                            System.out.println("member1 : " + member1);
-                            System.out.println("member2 : " + member2);
-                            System.out.println("member3 : " + member3);
 
                             // GeoPoint를 String으로 변환 (firebase -> spring)
                             GeoPoint startPoint = document.getGeoPoint("startPoint");
@@ -98,14 +99,8 @@ public class CarpoolService {
 
                 Long currentTime = System.currentTimeMillis(); // 현재 시간 (epoch 시간)
 
-                // 현재 시간(currentTime)과 carpoolResponseDTOS의 startTime을 비교
                 for (CarpoolResponseDTO carpoolDTO : carpoolResponseDTOS) {
-                    if ((currentTime - carpoolDTO.getStartTime()) / (24 * 60 * 60 * 1000) >= 90) { // 90일 이상이 지난 carpool
-
-//                        시간 계산 테스트
-//                        System.out.println("현재 시간 : " + currentTime);
-//                        System.out.println("carpool의 startTime : " + carpoolDTO.getStartTime());
-//                        System.out.println("차이 : " + (currentTime - carpoolDTO.getStartTime()) / (24 * 60 * 60 * 1000) + "일");
+                    if ((currentTime - carpoolDTO.getStartTime()) / (60 * 60 * 1000) >= 3) { // 3시간 이상이 지난 carpool
 
                         // currentTime에서 hour만 뽑기
                         Date date = new Date(currentTime);
@@ -113,12 +108,52 @@ public class CarpoolService {
                         String hours = sdf.format(date);
 
                         // 삭제할 carpool의 carId 토픽을 구독한 유저에게 푸시 알림 보내기
-                        fcmService.sendFcmMessage( hours + "시에 이용했던 카풀은 어떠셨나요?", "이용 내역에서 확인해보세요!", carpoolDTO.getCarId(), currentTime);
+                        fcmService.sendFcmMessage( hours + "시 " + carpoolDTO.getEndDetailPoint() + "까지의 카풀은 어떠셨나요?",
+                                "이용 내역에서 확인해보세요!",
+                                carpoolDTO.getCarId(), currentTime);
 
                         // carpool을 history로 옮기기
                         HistoryRequestDTO history = historyService.carpoolToHistory(carpoolDTO);
                         // spring에 carpool 저장
-                        historyService.saveHistory(history);
+                        try {
+                            historyService.saveHistory(history);
+                        } catch (IllegalArgumentException e) {
+                            // 이미 존재하는 carpoolId(이용내역)이 있을경우
+                        }
+
+                    }
+                }
+            } catch (ExecutionException | InterruptedException e) {
+                e.printStackTrace();
+            }
+
+        }, 3, 3600, TimeUnit.SECONDS); // 3초 후 부터 1시간마다 실행
+
+
+        // 출발시간으로부터 7일 이상 지난 firestore 카풀 삭제하는 스케줄링
+        scheduler.scheduleAtFixedRate(() -> {
+            try {
+
+                ApiFuture<QuerySnapshot> future = firestore.collection(COLLECTION_NAME).get();
+
+                List<QueryDocumentSnapshot> documents = future.get().getDocuments();
+
+                List<CarpoolShortDTO> carpoolShortDTOS = documents.stream()
+                        .map(document -> {
+                            CarpoolShortDTO carpool = new CarpoolShortDTO(
+                                    document.getString("carId"),
+                                    document.getString("endDetailPoint"),
+                                    document.getLong("startTime")
+                            );
+                            return carpool;
+                        })
+                        .collect(Collectors.toList());
+
+                Long currentTime = System.currentTimeMillis(); // 현재 시간 (epoch 시간)
+
+                // 현재 시간(currentTime)과 carpoolResponseDTOS의 startTime을 비교
+                for (CarpoolShortDTO carpoolDTO : carpoolShortDTOS) {
+                    if ((currentTime - carpoolDTO.getStartTime()) / (24 * 60 * 60 * 1000) >= 7) { // 7일 이상이 지난 carpool
 
                         // firestore에서 carpool을 삭제
                         firestore.collection(COLLECTION_NAME).document(carpoolDTO.getCarId()).delete();
@@ -132,6 +167,6 @@ public class CarpoolService {
             } catch (ExecutionException | InterruptedException e) {
                 e.printStackTrace();
             }
-        }, 5, 3600, TimeUnit.SECONDS); // 5초 후 부터 1시간마다 실행
+        }, 3, 3600, TimeUnit.SECONDS); // 3초 후 부터 1시간마다 실행
     }
 }
